==> 01_eda.ipynb <==
# Imports

import pandas as pd

# Load data

df = pd.read_csv('data/raw/cs-training.csv', index_col=0)
print(df.shape)
df.head(3)

# Health checks

df[df.duplicated(keep=False)].sort_values(by=list(df)).shape

Ми насправді не знаємо, чи дублікати це насправді, бо не маємо ключів

df.isna().sum()

Ми не знаємо природи пропусків тому будемо заповнювати по загальному підходу

# EDA

df.shape

df['SeriousDlqin2yrs'].value_counts() / len(df)

df.dtypes
