==> 12_rf_data_prep.ipynb <==
# Imports

import pandas as pd
from sklearn.model_selection import train_test_split


# Load data

df = pd.read_csv('data/raw/cs-training.csv', index_col=0)
print(df.shape)
df.head(3)

# Cut test and save

df_train, df_test = train_test_split(df, random_state=16, test_size=3000)
df_train.shape, df_test.shape

df_test.to_csv('data/raw/test.csv', index=False)

# Prepare data

## Train-val split

df_train, df_val = train_test_split(df_train, random_state=16, test_size=7000)
df_train.shape, df_val.shape

## Handle missing data

всі колонки числові. заповнимо медіанами + додамо колонки _is_NA

na_count = df_train.isna().sum()[df_train.isna().sum() != 0]
na_count

na_cols = list(na_count.index)
na_cols

mis_dict = {}

for c in na_cols:
    isna_train = df_train[c].isna()
    isna_val = df_val[c].isna()
    df_train[f'isna_{c}'] = True
    df_train.loc[isna_train, f'isna_{c}'] = True
    df_val[f'isna_{c}'] = True
    df_val.loc[isna_val, f'isna_{c}'] = True
    
    median_value = df_train[c].median()
    df_train.loc[isna_train, c] = median_value
    df_val.loc[isna_val, c] = median_value
    
    mis_dict[c] = median_value

df_train.isna().sum().sum(), df_val.isna().sum().sum()

import pickle as pkl

with open('models/missings.pkl', 'wb') as handle:
    pkl.dump(mis_dict, handle)

# Save data and models

df_train.to_csv('data/ready/train_rf.csv', index=False)

df_val.to_csv('data/ready/val_rf.csv', index=False)
