==> 13_rf.ipynb <==
# Imports

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import recall_score

# Load data

df_train = pd.read_csv('data/ready/train_rf.csv')
df_train.head()

df_val = pd.read_csv('data/ready/val_rf.csv')
df_val.head()

# Train model

cols_y = ['SeriousDlqin2yrs']
cols_x = list(set(df_train) - set(cols_y))
cols_x, cols_y

rf = RandomForestClassifier(n_estimators=1, n_jobs=1)
rf.fit(df_train[cols_x], df_train[cols_y])

y_train_pred = rf.predict(df_train[cols_x])
y_val_pred = rf.predict(df_val[cols_x])

train_recall = recall_score(df_train[cols_y], y_train_pred)
val_recall = recall_score(df_val[cols_y], y_val_pred)

print(f'Train recall: {train_recall}')
print(f'Validation recall: {val_recall}')

rf = RandomForestClassifier(n_jobs=5)
rf.fit(df_train[cols_x], df_train[cols_y])

y_train_pred = rf.predict(df_train[cols_x])
y_val_pred = rf.predict(df_val[cols_x])

train_recall = recall_score(df_train[cols_y], y_train_pred)
val_recall = recall_score(df_val[cols_y], y_val_pred)

print(f'Train recall: {train_recall}')
print(f'Validation recall: {val_recall}')

## min_samples_leaf

rf = RandomForestClassifier(n_jobs=5, min_samples_leaf=3)
rf.fit(df_train[cols_x], df_train[cols_y])

y_train_pred = rf.predict(df_train[cols_x])
y_val_pred = rf.predict(df_val[cols_x])

train_recall = recall_score(df_train[cols_y], y_train_pred)
val_recall = recall_score(df_val[cols_y], y_val_pred)

print(f'Train recall: {train_recall}')
print(f'Validation recall: {val_recall}')

rf = RandomForestClassifier(n_jobs=5, min_samples_leaf=5)
rf.fit(df_train[cols_x], df_train[cols_y])

y_train_pred = rf.predict(df_train[cols_x])
y_val_pred = rf.predict(df_val[cols_x])

train_recall = recall_score(df_train[cols_y], y_train_pred)
val_recall = recall_score(df_val[cols_y], y_val_pred)

print(f'Train recall: {train_recall}')
print(f'Validation recall: {val_recall}')

## max_features

rf = RandomForestClassifier(n_jobs=5, min_samples_leaf=1, max_features=0.5)
rf.fit(df_train[cols_x], df_train[cols_y])

y_train_pred = rf.predict(df_train[cols_x])
y_val_pred = rf.predict(df_val[cols_x])

train_recall = recall_score(df_train[cols_y], y_train_pred)
val_recall = recall_score(df_val[cols_y], y_val_pred)

print(f'Train recall: {train_recall}')
print(f'Validation recall: {val_recall}')

rf = RandomForestClassifier(n_jobs=5, min_samples_leaf=1, max_features=0.3)
rf.fit(df_train[cols_x], df_train[cols_y])

y_train_pred = rf.predict(df_train[cols_x])
y_val_pred = rf.predict(df_val[cols_x])

train_recall = recall_score(df_train[cols_y], y_train_pred)
val_recall = recall_score(df_val[cols_y], y_val_pred)

print(f'Train recall: {train_recall}')
print(f'Validation recall: {val_recall}')

rf = RandomForestClassifier(n_jobs=5, min_samples_leaf=1, max_features=0.75)
rf.fit(df_train[cols_x], df_train[cols_y])

y_train_pred = rf.predict(df_train[cols_x])
y_val_pred = rf.predict(df_val[cols_x])

train_recall = recall_score(df_train[cols_y], y_train_pred)
val_recall = recall_score(df_val[cols_y], y_val_pred)

print(f'Train recall: {train_recall}')
print(f'Validation recall: {val_recall}')

## Decision point optimization

rf = RandomForestClassifier(n_jobs=5, min_samples_leaf=1, max_features=0.75)
rf.fit(df_train[cols_x], df_train[cols_y])

decision_points = [0.01, 0.05] + [i * 0.1 for i in range(1, 10)]
scores = {}
for p in decision_points:
    
    y_train_pred = (rf.predict_proba(df_train[cols_x])[:, 1] > p).astype(int)
    y_val_pred = (rf.predict_proba(df_val[cols_x])[:, 1] > p).astype(int)
    display(y_train_pred)

    train_recall = recall_score(df_train[cols_y], y_train_pred)
    val_recall = recall_score(df_val[cols_y], y_val_pred)
    print(f'Decision point {p}')
    print(f'Train recall: {train_recall}')
    print(f'Validation recall: {val_recall}')
    scores[p] = (train_recall, val_recall)

scores

Best decision point is 0.01

# Save model

import pickle as pkl

model = {'model': rf
        , 'decision_point': 0.01}

with open('models/rf.pkl', 'wb') as handle:
    pkl.dump(model, handle)

# Predict and eval test

df_test = pd.read_csv('data/raw/test.csv')
df_test

df_test.isna().sum()

with open('models/missings.pkl', 'rb') as handle:
    miss_dict = pkl.load(handle)
    
miss_dict

for k in miss_dict.keys():
    df_test[f'isna_{k}'] = 0
    df_test.loc[df_test[k].isna(), f'isna_{k}'] = True
    df_test.loc[df_test[k].isna(), k] = miss_dict[k]
    
df_test.isna().sum()

y_test_pred = (rf.predict_proba(df_test[cols_x])[:, 1] > 0.01).astype(int)
test_recall = recall_score(df_test[cols_y], y_test_pred)

print(f'Test recall: {test_recall}')
