==> 21_mlp.ipynb <==
# Imports

import pandas as pd
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from tensorflow.keras.metrics import Recall
import tensorflow as tf

from sklearn.preprocessing import StandardScaler

import pickle as pkl

from sklearn.metrics import recall_score, accuracy_score

# Load data

df_train = pd.read_csv('data/ready/train_rf.csv')
df_train.head()

df_val = pd.read_csv('data/ready/val_rf.csv')
df_val.head()

# Scale

scaler = StandardScaler()

cols_x = list(set(df_train) - set(['SeriousDlqin2yrs']))

x_train = scaler.fit_transform(df_train[cols_x])

x_val = scaler.transform(df_val[cols_x])

# Model

class_weights = (df_train['SeriousDlqin2yrs'].value_counts() / len(df_train)).to_dict()

weight_0 = class_weights[0]
class_weights[0] = class_weights[1]
class_weights[1] = weight_0

class_weights

## Find good lr

model = Sequential()

model.add(
        Dense(100
          , input_shape=(len(cols_x),) 
          , activation='relu'
            )
         )

model.add(Dense(100, activation='relu'))
model.add(Dense(1, activation='sigmoid'))

recall = Recall()
model.compile(loss='binary_crossentropy',
              optimizer=tf.keras.optimizers.Adam(learning_rate=1e-2),
              metrics=[recall, 'accuracy'])

model.fit(x_train,df_train['SeriousDlqin2yrs'],epochs=10
          , class_weight=class_weights
          , validation_data=(x_val, df_val['SeriousDlqin2yrs']))

model = Sequential()

model.add(
        Dense(100
          , input_shape=(len(cols_x),) 
          , activation='relu'
            )
         )

model.add(Dense(100, activation='relu'))
model.add(Dense(1, activation='sigmoid'))

recall = Recall()
model.compile(loss='binary_crossentropy',
              optimizer=tf.keras.optimizers.Adam(learning_rate=1e-1),
              metrics=[recall, 'accuracy'])

model.fit(x_train,df_train['SeriousDlqin2yrs'],epochs=10
          , class_weight=class_weights
          , validation_data=(x_val, df_val['SeriousDlqin2yrs']))

model = Sequential()

model.add(
        Dense(100
          , input_shape=(len(cols_x),) 
          , activation='relu'
            )
         )

model.add(Dense(100, activation='relu'))
model.add(Dense(1, activation='sigmoid'))

recall = Recall()
model.compile(loss='binary_crossentropy',
              optimizer=tf.keras.optimizers.Adam(learning_rate=1e-3),
              metrics=[recall, 'accuracy'])

model.fit(x_train,df_train['SeriousDlqin2yrs'],epochs=10
          , class_weight=class_weights
          , validation_data=(x_val, df_val['SeriousDlqin2yrs']))

model.compile(loss='binary_crossentropy',
              optimizer=tf.keras.optimizers.Adam(learning_rate=1e-1),
              metrics=[recall])

model.fit(x_train,df_train['SeriousDlqin2yrs'],epochs=10
          , class_weight=class_weights
          , validation_data=(x_val, df_val['SeriousDlqin2yrs']))

1e-2 it is

## Bigger network

model = Sequential()

model.add(
        Dense(100
          , input_shape=(len(cols_x),) 
          , activation='relu'
            )
         )

model.add(Dense(100, activation='relu'))
model.add(Dense(100, activation='relu'))
model.add(Dense(1, activation='sigmoid'))

recall = Recall()
model.compile(loss='binary_crossentropy',
              optimizer=tf.keras.optimizers.Adam(learning_rate=1e-3),
              metrics=[recall, 'accuracy'])

model.fit(x_train,df_train['SeriousDlqin2yrs'],epochs=10
          , class_weight=class_weights
          , validation_data=(x_val, df_val['SeriousDlqin2yrs']))

model = Sequential()

model.add(
        Dense(100
          , input_shape=(len(cols_x),) 
          , activation='relu'
            )
         )

model.add(Dense(100, activation='relu'))
model.add(Dense(100, activation='relu'))
model.add(Dense(100, activation='relu'))
model.add(Dense(1, activation='sigmoid'))

recall = Recall()
model.compile(loss='binary_crossentropy',
              optimizer=tf.keras.optimizers.Adam(learning_rate=1e-3),
              metrics=[recall, 'accuracy'])

model.fit(x_train,df_train['SeriousDlqin2yrs'],epochs=10
          , class_weight=class_weights
          , validation_data=(x_val, df_val['SeriousDlqin2yrs']))

## Add batchNorm

model = Sequential()

model.add(
        Dense(200
          , input_shape=(len(cols_x),) 
          , activation='relu'
            )
         )

model.add(Dense(200, activation='relu'))
model.add(tf.keras.layers.BatchNormalization())
model.add(Dense(200, activation='relu'))
model.add(tf.keras.layers.BatchNormalization())
model.add(Dense(200, activation='relu'))
model.add(tf.keras.layers.BatchNormalization())
model.add(Dense(1, activation='sigmoid'))

recall = Recall()
model.compile(loss='binary_crossentropy',
              optimizer=tf.keras.optimizers.Adam(learning_rate=1e-3),
              metrics=[recall, 'accuracy'])

model.fit(x_train,df_train['SeriousDlqin2yrs'],epochs=10
          , class_weight=class_weights
          , validation_data=(x_val, df_val['SeriousDlqin2yrs']))

## Adjust class weights

class_weights = {0: 0.05, 1: 0.95}

model = Sequential()

model.add(
        Dense(200
          , input_shape=(len(cols_x),) 
          , activation='relu'
            )
         )

model.add(Dense(200, activation='relu'))
model.add(tf.keras.layers.BatchNormalization())
model.add(Dense(200, activation='relu'))
model.add(tf.keras.layers.BatchNormalization())
model.add(Dense(200, activation='relu'))
model.add(tf.keras.layers.BatchNormalization())
model.add(Dense(1, activation='sigmoid'))

recall = Recall()
model.compile(loss='binary_crossentropy',
              optimizer=tf.keras.optimizers.Adam(learning_rate=1e-3),
              metrics=[recall, 'accuracy'])

model.fit(x_train,df_train['SeriousDlqin2yrs'],epochs=10
          , class_weight=class_weights
          , validation_data=(x_val, df_val['SeriousDlqin2yrs']))

## More epochs

model = Sequential()

model.add(
        Dense(200
          , input_shape=(len(cols_x),) 
          , activation='relu'
            )
         )

model.add(Dense(200, activation='relu'))
model.add(tf.keras.layers.BatchNormalization())
model.add(Dense(200, activation='relu'))
model.add(tf.keras.layers.BatchNormalization())
model.add(Dense(200, activation='relu'))
model.add(tf.keras.layers.BatchNormalization())
model.add(Dense(1, activation='sigmoid'))

recall = Recall()
model.compile(loss='binary_crossentropy',
              optimizer=tf.keras.optimizers.Adam(learning_rate=1e-3),
              metrics=[recall, 'accuracy'])

model.fit(x_train,df_train['SeriousDlqin2yrs'],epochs=15
          , class_weight=class_weights
          , validation_data=(x_val, df_val['SeriousDlqin2yrs']))

model.save('models/keras_model')

# Test

df_test = pd.read_csv('data/raw/test.csv')
df_test.head()

with open('models/missings.pkl', 'rb') as handle:
    miss_dict = pkl.load(handle)

miss_dict

df_test.isna().sum()

for k in miss_dict.keys():
    df_test[f'isna_{k}'] = 0
    df_test.loc[df_test[k].isna(), f'isna_{k}'] = True
    df_test.loc[df_test[k].isna(), k] = miss_dict[k]
    
df_test.isna().sum()

x_test = scaler.transform(df_test[cols_x])

print('Train recall:',
    model.evaluate(x_train, df_train['SeriousDlqin2yrs'])[1]
     )

print('Validation recall:',
    model.evaluate(x_val, df_val['SeriousDlqin2yrs'])[1]
     )

print('Test recall:',
    model.evaluate(x_test, df_test['SeriousDlqin2yrs'])[1]
     )

# Save scaler

with open('models/nn_scaler.pkl', 'wb') as handle:
    pkl.dump(scaler, handle)
